+++ Source/Coco.Server/Coco.Entities/Domain/Dbo/Country.cs
@@ -1,5 +1,6 @@
 ﻿using Coco.Entities.Base;
 using Coco.Entities.Domain.Account;
+using System;
 using System.Collections.Generic;
 
 namespace Coco.Entities.Domain.Dbo
@@ -11,7 +12,7 @@ namespace Coco.Entities.Domain.Dbo
             this.UserInfos = new HashSet<UserInfo>();
         }
 
-        public int Id { get; set; }
+        public short Id { get; set; }
         public string Code { get; set; }
         public string Name { get; set; }
         public virtual ICollection<UserInfo> UserInfos { get; set; }
